SYNC-1285 Extract to new class
