+++ modules/util/sync-engine/src/com/liferay/sync/engine/service/SyncFileService.java
@@ -20,6 +20,7 @@ import com.liferay.sync.engine.model.ModelListener;
 import com.liferay.sync.engine.model.SyncFile;
 import com.liferay.sync.engine.model.SyncFileModelListener;
 import com.liferay.sync.engine.service.persistence.SyncFilePersistence;
+import com.liferay.sync.engine.util.FileKeyUtil;
 import com.liferay.sync.engine.util.FileUtil;
 import com.liferay.sync.engine.util.IODeltaUtil;
 
@@ -121,7 +122,7 @@ public class SyncFileService {
 
 		_syncFilePersistence.create(syncFile);
 
-		FileUtil.writeFileKey(
+		FileKeyUtil.writeFileKey(
 			Paths.get(filePathName), String.valueOf(syncFile.getSyncFileId()));
 
 		IODeltaUtil.checksums(syncFile);
@@ -569,7 +570,7 @@ public class SyncFileService {
 
 		// Local sync file
 
-		FileUtil.writeFileKey(
+		FileKeyUtil.writeFileKey(
 			filePath, String.valueOf(syncFile.getSyncFileId()));
 
 		Path deltaFilePath = null;
