+++ modules/util/sync-engine/src/com/liferay/sync/engine/util/FileUtil.java
@@ -14,21 +14,15 @@
 
 package com.liferay.sync.engine.util;
 
-import ch.securityvision.xattrj.Xattrj;
-
 import com.liferay.sync.engine.documentlibrary.util.FileEventUtil;
 import com.liferay.sync.engine.model.SyncFile;
 import com.liferay.sync.engine.service.SyncFileService;
 
-import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
 
-import java.nio.ByteBuffer;
-import java.nio.CharBuffer;
 import java.nio.channels.FileChannel;
 import java.nio.channels.FileLock;
-import java.nio.charset.Charset;
 import java.nio.file.FileSystem;
 import java.nio.file.FileSystems;
 import java.nio.file.FileVisitResult;
@@ -39,7 +33,6 @@ import java.nio.file.SimpleFileVisitor;
 import java.nio.file.StandardCopyOption;
 import java.nio.file.attribute.BasicFileAttributes;
 import java.nio.file.attribute.FileTime;
-import java.nio.file.attribute.UserDefinedFileAttributeView;
 
 import java.util.Arrays;
 import java.util.HashSet;
@@ -111,7 +104,7 @@ public class FileUtil {
 
 					if (syncFile == null) {
 						syncFile = SyncFileService.fetchSyncFile(
-							FileUtil.getFileKey(filePath));
+							FileKeyUtil.getFileKey(filePath));
 					}
 
 					if (syncFile != null) {
@@ -132,7 +125,7 @@ public class FileUtil {
 
 					if (syncFile == null) {
 						syncFile = SyncFileService.fetchSyncFile(
-							FileUtil.getFileKey(filePath));
+							FileKeyUtil.getFileKey(filePath));
 					}
 
 					if (syncFile != null) {
@@ -184,57 +177,6 @@ public class FileUtil {
 		}
 		finally {
 			StreamUtil.cleanUp(fileInputStream);
-		}
-	}
-
-	public static long getFileKey(Path filePath) {
-		if (!Files.exists(filePath)) {
-			return -1;
-		}
-
-		try {
-			if (OSDetector.isApple()) {
-				Xattrj xattrj = getXattrj();
-
-				if (xattrj == null) {
-					return -1;
-				}
-
-				String fileKey = xattrj.readAttribute(
-					filePath.toFile(), "fileKey");
-
-				if (fileKey == null) {
-					return -1;
-				}
-
-				return Long.parseLong(fileKey);
-			}
-			else {
-				UserDefinedFileAttributeView userDefinedFileAttributeView =
-					Files.getFileAttributeView(
-						filePath, UserDefinedFileAttributeView.class);
-
-				List<String> list = userDefinedFileAttributeView.list();
-
-				if (!list.contains("fileKey")) {
-					return -1;
-				}
-
-				ByteBuffer byteBuffer = ByteBuffer.allocate(
-					userDefinedFileAttributeView.size("fileKey"));
-
-				userDefinedFileAttributeView.read("fileKey", byteBuffer);
-
-				CharBuffer charBuffer = _CHARSET.decode(
-					(ByteBuffer)byteBuffer.flip());
-
-				return Long.parseLong(charBuffer.toString());
-			}
-		}
-		catch (Exception e) {
-			_logger.error(e.getMessage(), e);
-
-			return -1;
 		}
 	}
 
@@ -406,7 +348,8 @@ public class FileUtil {
 			}
 
 			if ((fileTime.toMillis() <= modifiedTime) &&
-				(getFileKey(filePath) == syncFile.getSyncFileId())) {
+				(FileKeyUtil.getFileKey(filePath) ==
+					syncFile.getSyncFileId())) {
 
 				return false;
 			}
@@ -551,31 +494,6 @@ public class FileUtil {
 		Files.setLastModifiedTime(filePath, fileTime);
 	}
 
-	public static void writeFileKey(final Path filePath, final String fileKey) {
-		if (FileUtil.getFileKey(filePath) == Long.parseLong(fileKey)) {
-			return;
-		}
-
-		if (Files.isDirectory(filePath)) {
-			doWriteFileKey(filePath, fileKey);
-
-			return;
-		}
-
-		PathCallable pathCallable = new PathCallable(filePath) {
-
-			@Override
-			public Object call() throws Exception {
-				doWriteFileKey(filePath, fileKey);
-
-				return null;
-			}
-
-		};
-
-		FileLockRetryUtil.registerPathCallable(pathCallable);
-	}
-
 	protected static void checkFilePath(Path filePath) {
 
 		// Check to see if the file or folder is still being written to. If
@@ -600,71 +518,10 @@ public class FileUtil {
 		}
 	}
 
-	protected static void doWriteFileKey(Path filePath, String fileKey) {
-		if (FileUtil.getFileKey(filePath) == Long.parseLong(fileKey)) {
-			return;
-		}
-
-		if (OSDetector.isApple()) {
-			Xattrj xattrj = getXattrj();
-
-			if (xattrj == null) {
-				return;
-			}
-
-			File file = filePath.toFile();
-
-			if (!file.canWrite()) {
-				file.setWritable(true);
-			}
-
-			xattrj.writeAttribute(file, "fileKey", fileKey);
-		}
-		else {
-			File file = filePath.toFile();
-
-			if (!file.canWrite()) {
-				file.setWritable(true);
-			}
-
-			UserDefinedFileAttributeView userDefinedFileAttributeView =
-				Files.getFileAttributeView(
-					filePath, UserDefinedFileAttributeView.class);
-
-			try {
-				userDefinedFileAttributeView.write(
-					"fileKey", _CHARSET.encode(CharBuffer.wrap(fileKey)));
-			}
-			catch (Exception e) {
-				_logger.error(e.getMessage(), e);
-			}
-		}
-	}
-
-	protected static Xattrj getXattrj() {
-		if (_xattrj != null) {
-			return _xattrj;
-		}
-
-		try {
-			_xattrj = new Xattrj();
-
-			return _xattrj;
-		}
-		catch (IOException ioe) {
-			_logger.error(ioe.getMessage(), ioe);
-
-			return null;
-		}
-	}
-
-	private static final Charset _CHARSET = Charset.forName("UTF-8");
-
 	private static final Logger _logger = LoggerFactory.getLogger(
 		FileUtil.class);
 
 	private static final Set<String> _syncFileIgnoreNames = new HashSet<>(
 		Arrays.asList(PropsValues.SYNC_FILE_IGNORE_NAMES));
-	private static Xattrj _xattrj;
 
 }
